fix type

+++ main.py
@@ -24,7 +24,7 @@ parser.add_argument('--train', type=bool, default=True, help="if true then train
 parser.add_argument('--test', type=bool, default=True, help="if true then test is done")
 parser.add_argument('--epochs', type=int, default=100, help="number of epoch")
 parser.add_argument('--save_weights', type=bool, default=False, help="if true then the weights are saved in each epoch")
-parser.add_argument('--adversarial_bound', type=int, default=0,
+parser.add_argument('--adversarial_bound', type=float, default=0,
                     help="if non-zero then the training is done using adversarial attack, where epsilon is the given value")
